backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.15.0
neo4j>=4.4.0
python-dotenv>=0.19.0
pydantic>=2.0.0
loguru>=0.5.3
selenium>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
websockets
webdriver-manager>=3.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
redis>=4.6.0
async-timeout>=4.0.0

# Monitoring and metrics
prometheus-client>=0.17.0

# ML and AI dependencies
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# XAI (Explainable AI) libraries - опциональные
# shap>=0.42.0
# lime>=0.2.0

# OpenAI integration - заменено на универсальный адаптер
# openai>=1.3.0  # Заменено на наш универсальный адаптер
httpx>=0.24.0  # Для HTTP запросов в адаптере
requests>=2.31.0  # Для REST API вызовов

# Anthropic Claude integration - опционально
# anthropic>=0.7.0

# Additional ML utilities
joblib>=1.3.0
onnx>=1.14.0
onnxruntime>=1.15.0
